Fall off Chasm
On Map003, I have adjusted the level to be more like Faiza's vision. The path tiles are walkable, even when they aren't lit up by the lightbulbs. And, if the player walks onto a chasm tile that is not the path, they will be teleported to the beginning of the map.