Chasm03 Respawned
Some of the 'Tricky Tiles' have been respawned to areas that were closer to where the player previously left off (to decrease player frustration).

I have not included how the black areas will respawn the player back to the beginning. Thought it better to test and find bugs before adding that step.