Light&Shadow Puzzle - Demo Implementation
Have completed demo of the Light&Shadow Puzzle Tutorial. Used tree as both the object being pushed around, and the shadow of it. Able to push the object around the sandy area, but not outside the bounds of this sandy area.
If the tree is pushed to the left or the right, a blue door will be revealed, which can be used to exit the level back into the lobby. If the tree is pushed back into the same column of the door, it will "dissapear" behind the shadow of the tree.