Skeleton Draft
Using the previously named P3 Room 1 as a skeleton room that contains all the parts that will make up the next few levels.

Started creating the skeleton. So far it contains:
(1) The 3 buttons corresponding to the 3 different colours.
(2) There are multiple lightbulbs of the same colour that have differing intensities and shapes.